Move session recording under extensions
